fix(Profiel): Style all other profile buttons
