+++ src/app/Components/NextUIBasedButton.tsx
@@ -18,8 +18,8 @@ export const NextUIBasedButton: React.FC<ExtraButtonProps> = ({
 }) => {
     const variantClass = {
         primary: 'bg-primary-100 hover:bg-primary-200 text-white h-12 px-7',
-        secondary: 'bg-gray-200 text-black hover:bg-gray-300',
-        profile: 'bg-transparant hover:text-orange-600 font-semibold hover:'
+        secondary: 'bg-gray-200 text-black hover:bg-gray-300 h-12 px-7',
+        profile: 'bg-transparant hover:text-orange-600 font-semibold'
     };
 
     return (
